+++ apex/LAMMPS_OPs.py
@@ -12,8 +12,8 @@ try:
     from monty.serialization import loadfn
     from dflow.python import upload_packages
     from apex.lib.utils import return_prop_list
-    from apex.property.common_equi import (make_equi, post_equi)
-    from apex.property.common_prop import (make_property, post_property)
+    from apex.core.common_equi import (make_equi, post_equi)
+    from apex.core.common_prop import (make_property, post_property)
     upload_packages.append(__file__)
 except:
     pass
@@ -45,7 +45,7 @@ class RelaxMakeLAMMPS(OP):
             self,
             op_in: OPIO,
     ) -> OPIO:
-        from apex.property.common_equi import make_equi
+        from apex.core.common_equi import make_equi
 
         cwd = os.getcwd()
         os.chdir(op_in["input"])
@@ -145,7 +145,7 @@ class RelaxPostLAMMPS(OP):
 
     @OP.exec_sign_check
     def execute(self, op_in: OPIO) -> OPIO:
-        from apex.property.common_equi import post_equi
+        from apex.core.common_equi import post_equi
 
         cwd = os.getcwd()
         os.chdir(str(op_in['input_all'])+op_in['path'])
@@ -207,7 +207,7 @@ class PropsMakeLAMMPS(OP):
             self,
             op_in: OPIO,
     ) -> OPIO:
-        from apex.property.common_prop import make_property
+        from apex.core.common_prop import make_property
 
         cwd = os.getcwd()
         os.chdir(op_in["input"])
@@ -275,7 +275,7 @@ class PropsPostLAMMPS(OP):
 
     @OP.exec_sign_check
     def execute(self, op_in: OPIO) -> OPIO:
-        from apex.property.common_prop import post_property
+        from apex.core.common_prop import post_property
 
         cwd = os.getcwd()
         os.chdir(str(op_in['input_all'])+op_in['path'])
