+++ apex/core/lib/trans_tools.py
