+++ apex/core/template/elastic/lmp/init.mod
