+++ apex/core/template/elastic/lmp/potential.mod
