+++ apex/op/relaxation_op.py
@@ -11,7 +11,7 @@ from dflow.python import (
 try:
     from monty.serialization import loadfn
     from apex.lib.utils import return_prop_list
-    from apex.property.common_equi import (make_equi, post_equi)
+    from apex.core.common_equi import (make_equi, post_equi)
 except:
     pass
 
@@ -46,7 +46,7 @@ class RelaxMake(OP):
             self,
             op_in: OPIO,
     ) -> OPIO:
-        from apex.property.common_equi import make_equi
+        from apex.core.common_equi import make_equi
 
         cwd = os.getcwd()
         os.chdir(op_in["input"])
@@ -89,7 +89,7 @@ class RelaxMake(OP):
 
 class RelaxPost(OP):
     """
-    class for analyzing calculation results
+    OP class for analyzing calculation results
     """
 
     def __init__(self):
@@ -113,7 +113,7 @@ class RelaxPost(OP):
 
     @OP.exec_sign_check
     def execute(self, op_in: OPIO) -> OPIO:
-        from apex.property.common_equi import post_equi
+        from apex.core.common_equi import post_equi
 
         cwd = os.getcwd()
         os.chdir(str(op_in['input_all']) + op_in['path'])
