+++ tests/test_elastic.py
@@ -1,23 +1,16 @@
 import glob
-import json
 import os
 import shutil
 import sys
 import unittest
 
-import dpdata
-import numpy as np
-from monty.serialization import dumpfn, loadfn
-from pymatgen.analysis.elasticity.strain import Deformation, Strain
 from pymatgen.core import Structure
 from pymatgen.io.vasp import Incar
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 __package__ = "tests"
 
-from apex.property.Elastic import Elastic
-
-from .context import make_kspacing_kpoints, setUpModule
+from apex.core.property import Elastic
 
 
 class TestElastic(unittest.TestCase):
