+++ tests/test_eos.py
@@ -1,5 +1,4 @@
 import glob
-import json
 import os
 import shutil
 import sys
@@ -7,15 +6,13 @@ import unittest
 
 import dpdata
 import numpy as np
-from monty.serialization import dumpfn, loadfn
+from monty.serialization import loadfn
 from pymatgen.io.vasp import Incar
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 __package__ = "tests"
 
-from apex.property.EOS import EOS
-
-from .context import make_kspacing_kpoints, setUpModule
+from apex.core.property import EOS
 
 
 class TestEOS(unittest.TestCase):
