+++ tests/test_interstitial.py
@@ -1,24 +1,18 @@
 import glob
-import json
 import os
 import shutil
 import sys
 import unittest
 
-import dpdata
 import numpy as np
-from monty.serialization import dumpfn, loadfn
 from pymatgen.analysis.defects.core import Interstitial as pmg_Interstitial
 from pymatgen.core import Structure
 from pymatgen.io.vasp import Incar
-from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 __package__ = "tests"
 
-from apex.property.Interstitial import Interstitial
-
-from .context import make_kspacing_kpoints, setUpModule
+from apex.core.property.Interstitial import Interstitial
 
 
 class TestInterstitial(unittest.TestCase):
