+++ tests/test_mpdb.py
@@ -8,7 +8,7 @@ from pymatgen.core import Structure
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 __package__ = "tests"
-from apex.property.mpdb import get_structure
+from apex.core.mpdb import get_structure
 
 from .context import make_kspacing_kpoints, setUpModule
