+++ tests/test_surface.py
@@ -1,13 +1,10 @@
 import glob
-import json
 import os
 import shutil
 import sys
 import unittest
 
-import dpdata
-import numpy as np
-from monty.serialization import dumpfn, loadfn
+from monty.serialization import loadfn
 from pymatgen.core import Structure
 from pymatgen.core.surface import SlabGenerator
 from pymatgen.io.vasp import Incar
@@ -15,9 +12,7 @@ from pymatgen.io.vasp import Incar
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 __package__ = "tests"
 
-from apex.property.Surface import Surface
-
-from .context import make_kspacing_kpoints, setUpModule
+from apex.core.property.Surface import Surface
 
 
 class TestSurface(unittest.TestCase):
