+++ tests/test_vacancy.py
@@ -1,13 +1,10 @@
 import glob
-import json
 import os
 import shutil
 import sys
 import unittest
 
-import dpdata
 import numpy as np
-from monty.serialization import dumpfn, loadfn
 from pymatgen.analysis.defects.core import Vacancy as pmg_Vacancy
 from pymatgen.core import Structure
 from pymatgen.io.vasp import Incar
@@ -16,9 +13,7 @@ from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 __package__ = "tests"
 
-from apex.property.Vacancy import Vacancy
-
-from .context import make_kspacing_kpoints, setUpModule
+from apex.core.property.Vacancy import Vacancy
 
 
 class TestVacancy(unittest.TestCase):
